docs: Highlight reader's position in table of contents
While browsing the documentation, I often become disoriented because
the table of contents is so long, and there's no connection between it
and the content I'm viewing.

To make it easier for readers to orient themselves, navigation links in
the table of contents are emphasized when their respective section is
being viewed by the user.

To accomplish this, `<section>` tags have been wrapped around all the
content. (I also took the liberty of using `<article>` and `<nav>`
where it was appropriate.) When the page's DOM is loaded, some
JavaScript runs that:

1. Creates an [`IntersectionObserver`][1].
2. Queries all the navigation links in the table of contents.
3. For each link, associates it with the destination section.
4. Passes each section to the `IntersectionObserver`.

When a section moves in or out of the viewport, a callback adds or
removes a `visible-section` class to the associated navigation link.

Known Issues:

- To be consistent, `header_open|close` should probably be renamed to
  `section_open|close`. I have not done that yet since it would bloat
  the diff and make the functional changes harder to review.
- Chrome seems to shrink headers right after a `<section>`. It looks
  a bit odd, so some CSS might be needed.

[1]: https://developer.mozilla.org/en-US/docs/Web/API/IntersectionObserver