Add tests to formatted printing of pointers
One of the purposes of these new tests is to validate that code that
would have previously executed paths that are now @compileErrors do not
execute those paths.

Rename the tests of have a more consistent style allowing easier
use of the test-filter parameter.

Add factor testExpectedActual out of testFmt and make it public so this
and other tests may use it.