Change convertSmallInt to receive the array rather than a byte.
This fixes an index out of bounds bug, as the byte array passed can be
empty when reading a u0. I also changed the u0 tests to pass an empty
buffer to test that case.

Also, added a test for a type u9 to validate that bit level reads work for
types with > 8 and not multiple of 8 bits.

Josh Wolfe has pointed out, https://github.com/ziglang/zig/pull/1569#issuecomment-423402695,
that these routines will not handle using the same buffer for subsequent
non-byte aligned reads. That is certainly true, but this API is not
suitable for that type of operation as it would require state to be
maintained between calls.

But these functions do handle reading the next N bits from the beginning
of the buffer, which appears to be all that is expected.

I found this problem because DefaultPrng.random.scalar was not working
for u0..u7 types and it uses readInt which it expects to handle any type.
DefaultPrng.random.range uses readInt too, with the same expectations.

So I don't believe this PR changes the semantics of the functions, it
simply allows all TypeId.Ints to be used, not just those whose length
in bits is >= 8.