copy elision: move the first behavior test to copy elision
This commit begins a process of migrating from

 * test/behavior.zig
 * test/cases/*

to

 * test/stage1/behavior.zig
 * test/stage1/behavior/*

Each test that this copy-elision branch is able to pass
will be moved to the new location. Once all tests pass,
it's time to merge into master.

This commit simplifies the default panic handler and
bootstrap code to disable stack traces, since those are
not yet working in the copy-elision branch. Once that code
works again, they can be re-enabled. That must happen before
merging into master.