Add vendor/llvm
Add LLVM as a submodule to zig so that LLVM maybe easily compiled and
different versions of LLVM can be tested as LLVM evolves.

The acutal submodule is vendo/llvm/src which defaults to LLVM-8.0.0 and
vendor/llvm contain a Makefile that allows it to be compiled as simply as
issuing a `make -jN` command. But also allows different versions to be
created by including Makefiles (aka .cfg files) which set variables that
control the build process.

Adds LLVM_LINK_STATIC as a cmake variable and suggests linking
vendor/llvm statically if LLVM isn't found or is the wrong version.
This avoids the need to set RPATH or LD_LIBRARY_PATH.

Modify README.md adding instructions for building LLVM using the
submodule. Also add a link to `#build-llvm` in README.md in error
messages in cmake/Findllvm.cmake.