Move FailingAllocator deallocation count increment
The mem.Allocator interface was changed in 9c13e9b. FailingAllocator
(along with other implementations) was updated accordingly, but its
`deallocations` increment was removed from what is now `shrinkFn`. The
result is that `index` and `deallocations` are always equal, both
incremented inside `reallocFn`, and no deallocations are recorded on
calls to `shrinkFn`.

Note that currently, the only reference to FailingAllocator's
`deallocations` field in the standard library is in
std/zig/parser_test.zig.