cmake: rework LLVM/Clang cmake
1. Use {LLVM,CLANG}_INCLUDE_DIRS from {LLVM,Clang}Config.cmake
  This makes it possible to use an LLVM/Clang source tree.

2. Delete ZIG_STATIC_LLVM

Building both libLLVM*.a and libLLVM*.so is unsupported in LLVM. The
LLVM libraries can be configured in 3 ways:

* -DLLVM_LINK_LLVM_DYLIB=ON => link against libLLVM.so
* -DLLVM_LINK_LLVM_DYLIB=OFF -DBUILD_SHARED_LIBS=OFF => link against libLLVM*.a
* -DLLVM_LINK_LLVM_DYLIB=OFF -DBUILD_SHARED_LIBS=ON => link against libLLVM*.so

If libLLVM*.a are unavailable, ZIG_STATIC_LLVM will not work.
If libLLVM*.a are available, linking against libLLVM*.a is automatic.
So ZIG_STATIC_LLVM is not useful and can only be confusing.

3. If CLANG_LINK_CLANG_DYLIB (available since clang 9) is ON,
libclang-cpp.so instead of libclang*.so will be linked. It currently
does not work because zig_cpp/libembedded_lld_lib.a has global
constructors that conflict with libclang-cpp.so